split SCSS files
