+++ package.json
@@ -24,6 +24,7 @@
     "babel-core": "~5.5.6",
     "babel-jest": "~5.3.0",
     "babel-loader": "~5.1.4",
+    "file-loader": "git+https://github.com/webpack/file-loader.git",
     "jest-cli": "~0.4.13",
     "json-loader": "^0.5.2",
     "node-libs-browser": "~0.5.2",
