+++ src/assets/img/index.js
@@ -1,0 +1,5 @@
+export { default as background } from './background.jpg'
+export { default as swords } from './crossed-swords.svg'
+export { default as galaxyShadow } from './galaxy-color-shadow.svg'
+export { default as leather } from './leather-tex.jpg'
+export { default as shield } from './shield.svg'
