+++ src/js/views/gameview/_gameview.scss
@@ -1,0 +1,116 @@
+@import "../../../assets/styles/base";
+
+/* 
+   styles for play area 
+*/
+
+// the overall container with all elements
+.game-container {
+  text-align: center; 
+  background-color: $palette-four; 
+}
+
+// area for cards to be played in center
+.in-play-area {
+  text-align: center; 
+  background-color:$palette-four; 
+  height: $in-play-height;
+  width: $in-play-width;
+  border-radius: $default-corner-radius;
+  margin-left: 16vw;
+  margin-bottom: $default-vertical-padding*-4;
+}
+
+// a player's in-play cards
+.player-board {
+  clear:both;
+  width:100%;
+  border: 1px solid;
+  background-color: rgba(245,245,245,1);
+  height: $card-height;
+  border-radius: $default-corner-radius;
+}
+
+// the player's board who is on bottom
+.home-player-board {
+  position: relative;
+  bottom: -10px;
+}
+
+// where a player's unplayed cards go
+.hand-container {
+  height:$card-height + $default-vertical-padding*4;
+  position: relative;
+  right: $end-turn-button-width/2;
+}
+
+/* 
+   styles for player avatar 
+*/
+
+// the representation of the player
+.player-avatar {
+  padding: 10px;
+  padding-top: 0; 
+  margin-bottom: $default-padding;
+  width: 120px;
+  height: 120px;
+  margin-left: 43vw;
+  font-weight: fontWeight; 
+  border-radius: 60px;
+  background-color: rgba(245,245,245,1);
+}
+
+// the bottom player
+.home-avatar {
+  margin-top:30px;
+}
+
+// life and mana
+.vital-stats-container {
+  color:black;
+  padding-top:1vh;
+}
+
+// life of player
+.life-container {
+  padding-top:2vh;
+}
+
+
+/* 
+   styles for end turn button 
+*/
+
+// the main End Turn button that always shows in the game
+.end-turn-button, .resign-button {
+  width: $end-turn-button-width;
+  padding:$default-padding;
+  border-radius:6px;
+  border: 1px solid;
+  position:relative;
+  bottom:$in-play-height/2 - 1vh;
+  margin-left: $in-play-width + 17vw + 1vw;
+}
+
+.end-turn-button {
+  // it's never time to end the turn in spacetime!
+  display: none; 
+  pointer-events:none;
+  background-color: $palette-four;
+  color:gray;
+}
+
+// highlight end of turn button for active player
+.active-player {
+  background-color: $palette-two;
+  box-shadow: 5px 5px 12px 0px rgba(22,22,22,1);
+  color:white;
+  cursor: pointer;
+  pointer-events:auto;
+}
+
+// pop up end turn button a bit on hover
+.active-player:hover {
+  box-shadow: 7px 7px 17px 0px rgba(22,22,22,1);
+}
