+++ src/js/views/lobbyview/_lobbyview.scss
@@ -1,0 +1,39 @@
+@import "../../../assets/styles/base";
+
+/* 
+   styles for LobbyView 
+*/
+
+// the overall container for the main lobby
+.lobby-container {
+  width: 94vh;
+  height:94vh;
+  text-align: center; 
+  background-color: rgba(0,0,0,.8); 
+  margin: 0 auto;
+  border-radius: 1vh;
+  border: 10px double #000; 
+  border-width: 1.5vh;
+  color:white;
+}
+
+// the button to find a game in the lobby
+.find-game-button {
+  width: $end-turn-button-width;
+  padding:$default-padding;
+  margin: 0 auto;
+  background-color: $palette-four;
+  border-radius:6px;
+  border: 1px solid;
+}
+
+.campaign-button {
+  margin-top:2vh;
+  width: $end-turn-button-width*2;
+}
+
+// the style for the big swords in the lobby
+.home-image {
+  margin-top:30px;
+  width:70%;
+}
