+++ src/js/views/lobbyview/lobbyview.js
@@ -1,7 +1,9 @@
 // React view of the Lobby
 import React from 'react';
-import Client from "./client"
-import GameState from "./gamestate"
+import Client from "../../client"
+import GameState from "../../gamestate"
+import "./_lobbyview.scss"
+import { galaxyShadow } from '../../../assets/img'
 
 let LobbyView = React.createClass({
     render() {
@@ -15,7 +17,7 @@ let LobbyView = React.createClass({
           <div className="find-game-button campaign-button" onClick={this.playCampaign}>
             Campaign 1: Bibot Attack
           </div>
-          <img className="home-image" src="img/galaxy-color-shadow.svg" />
+          <img className="home-image" src={galaxyShadow} />
         </div>
     );
   },
