+++ webpack.config.js
@@ -6,7 +6,7 @@ var csswring     = require('csswring');
 module.exports = {  
   entry: [
     'webpack/hot/only-dev-server',
-    "./js/app.js"
+    "./src/js/app.js"
   ],
   output: {
     path: __dirname + '/build',
@@ -16,13 +16,17 @@ module.exports = {
     loaders: [
       { test: /\.js?$/, loaders: ['react-hot', 'babel'], exclude: /node_modules/ },
       { test: /\.js$/, exclude: /node_modules/, loader: 'babel-loader'},
-      { test: /\.scss$/, loader: "style!css!postcss-loader!sass" }    
+      { test: /\.scss$/, loader: "style!css!postcss-loader!sass" },
+      { test: /\.(jpe?g|png|gif|svg|ico|cur)$/i, loader: 'file-loader?name=images/[name].[ext]'}
 
     ]
   },
   plugins: [
     new webpack.NoErrorsPlugin()
   ],
+  resolve: {
+    extensions: ['', '.js']
+  },
   postcss: function () {
     return [autoprefixer, csswring]
   }
